buffers: handle very short fragmentation of handshake messages
If the received record doesn't even complete the handshake
header (i.e., the record size < 4), keep it in a temporary buffer and
let the caller receive more records.  Once enough amount of data is
received, move the already received records back to record_buffer and
proceed to the normal processing.

Fixes: #272

Signed-off-by: Daiki Ueno <dueno@redhat.com>