_gnutls_server_select_cert: return error when no server cert is selected
When a certificate callback is used and no certificate is provided
by it, return an error rather than trying to use it (and crashing)
later. Note that this affects only an "illegal" code path when
a server would have provided no certificate, something which must
not happen on a real-world server.

Signed-off-by: Nikos Mavrogiannopoulos <nmav@redhat.com>