tests: tpm: Add a test case for tpmtool
This test case exercises tpmtool and uses certtool to create a
self-signed certificate with the TPM. It uses swtpm as TPM emulator and
configures tcsd to talk to swtpm.

Extend the Readme.md with the packages needed for TPM support and TPM test
support.

This test case needs to be run as root since tcsd needs to be started
as root.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>