Prevent applications from combining legacy versions of TLS with TLS1.3
It can happen that an application due to a misconfiguration, enables TLS1.3
in combination with TLS1.0 or TLS1.1 only. In that case a server which is
unaware of the TLS1.3 protocol will reply by selecting the TLS1.2 protocol
instead and that answer will be rejected by the client. With this change
we ensure that TLS1.3 is not enabled in these problematic scenarios.

Resolves: #621

Signed-off-by: Nikos Mavrogiannopoulos <nmav@redhat.com>