bootstrap: refuse to bootstrap if any dependencies bring gnulib's network stack
If gnulib's network stack is brought (due to a dependency) in the library
it will make the library unusable to non-gnulib using applications. This
prevents windows applications for example to use gnutls, and so on. Even
more it is quite hard to catch that issue because our testsuite uses
gnulib as well. Instead we try to catch the these modules at import time.

Signed-off-by: Nikos Mavrogiannopoulos <nmav@gnutls.org>