gnutls-cli: Add option "--logfile" to redirect information message output
First, add an option "--logfile" so user could choose a specific file to
store all the informational messages. In some cases, informational
messages may cause unexpected result if the output is standard output.

With this option, user could redirect these messages to a specific
file. This will be helpful in testing and tracking.

Second, replace printf() function with log_msg() function

This log_msg() function is used when "--logfile" is enabled.

Third, add a functionality test for "--logfile" option

Add a test script to test if "--logfile" option works as it should be.

Signed-off-by: Ke Zhao <kzhao@redhat.com>