Fix check_if_signed
Fix the target branch we check against by adding upstream as remote.

Drop the use of set -e as this causes the shell to immediately exit on
errors instead of allowing the code to check the failure and report what
it faled about.

Also print which commits are being checked and what information was found
so that a CI failure can be better diagnosed.

Signed-off-by: Simo Sorce <simo@redhat.com>