crypto-selftests-pk.c: Use deterministic signatures in test_known_sig()
Use deterministic signatures for ECDSA and DSA in test_known_sig().  Do
not call test_known_sig() for non-deterministic algorithms.  Do not run
PK_TEST() for algorithms tested with PK_KNOWN_TEST().

The deterministic algorithms are used if in FIPS-140 POST or if FIPS-140
mode is disabled.  When called explicitly with FIPS-140 mode enabled,
the pairwise-consistency test (PK_TEST()) is used instead.

test_known_sig() was modified to support only deterministic algorithms.
The "deterministic" parameter was replaced with the "flags" parameter
through which the flags to be used in gnutls_privkey_sign_data() are
passed.

The hard-coded values for the ECDSA and DSA signatures were replaced
with the values corresponding to the deterministic signatures to be used
in known answer tests.  The unused values for GOST signatures were
removed.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>