crypto-selftests-pk.c: Fix PK_KNOWN_TEST and PK_TEST
Previously, when multiple tests where declared in sequence using one of
the macros, only the first test would be executed.  This happened
because a check for the GNUTLS_SELF_TEST_FLAG_ALL was embedded in the
macro.  To allow more than one test to be executed in sequence, the
check for the flag was removed from both macros.

To keep the previous behaviour (execute only the first test) the check
for the flag was moved to be after the first test, except for RSA since
the RSA encryption test must be executed in FIPS mode.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>