crypto-selftests-pk.c: Fix PK_KNOWN_TEST and PK_TEST
Remove the flag check from the end of the macros.  This change allows
more than one test to run in sequence when GNUTLS_SELF_TEST_FLAG_ALL is
not set.  Move the flags checks to run the minimal set of tests required
for FIPS and keep the previous behaviour for GOST (run the first test
for each algorithm).

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>