tests: fix race condition in tls13/post-handshake-with-cert-pkcs11
The test had a strange setup of server/client processes: the server
runs in a child process and the client runs in a parent process.  The
intention behind this was to detect softhsm availability in the parent
process and exit with 77 if missing.  However, there was a potential
race when the server exits and proceeds to the next call of start().

This fixes the process setup and moves the softhsm detection at the
program startup.

Signed-off-by: Daiki Ueno <dueno@redhat.com>