certtool: generate RSA-PSS certificates from RSA keys
When generating certificates it was not possible to generate
an RSA-PSS certificate from an RSA key (common scenario). This
fixes the certificate generation to include such a method.

Ironically there was a test for this scenario but the test
was limited to checking that the combination of certtool parameters
succeeded; modified the test to check the textual expression of
the certificate for the RSA-PSS indicators.

Signed-off-by: Nikos Mavrogiannopoulos <nmav@redhat.com>