tests: Run rng-no-onload test in FIPS mode
This changes the function used in the test to override gnutls_rnd() to
fill the given buffer with a different value each time it is called.
This allows the test to run when FIPS mode is enabled.

Previously the rng-no-onload test could get stuck if FIPS mode was
enabled.  This happened if gnutls_rnd() function was called during
global_init() in a loop that checks the generated value (e.g. if ECDSA
signature generation is called during self tests).

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>