.gitlab-ci.yml: removed coverity build [ci skip]
The coverity run is subject to several restrictions by the service,
and thus it is not really useful in the main CI runs as it cannot reasonably
be run on MRs or master. As such we simplify the main CI file by moving the
coverity to the coverage sub-project and running it weekly.

The new location is at:
https://gitlab.com/gnutls/coverage

Signed-off-by: Nikos Mavrogiannopoulos <nmav@redhat.com>