lib: use static assertion to check enum values
We previously had checks of enum values with '#if', such as below:

 #define GNUTLS_EXTENSION_MAX_VALUE 31

 typedef enum extensions_t {
         ...
         GNUTLS_EXTENSION_MAX /* not real extension - used for iterators */
 } extensions_t;

 /* we must provide at least 16 extensions for users to register */
 #if GNUTLS_EXTENSION_MAX_VALUE - GNUTLS_EXTENSION_MAX < 16
 # error not enough extension types
 #endif

This doesn't work as expected; because GNUTLS_EXTENSION_MAX is not
defined as a preprocessor macro, it always expands to 0.  To properly
do this check, we need to use static assert as provided as the
'verify' macro in gnulib.

Signed-off-by: Daiki Ueno <dueno@redhat.com>