x509: Improve documentation of new set_getissuer_function
Since gnutls!1354, some of this information is now obsolete. The caller
is no longer responsible for verifying the certificate or adding it to
the trust list. GnuTLS will now handle that. Instead, the callback
should always import the missing certificate and return success if the
certificate was imported, or failure otherwise.

Also, let's point to gnutls_x509_crt_get_authority_info_access(), since
it is useful in combination with this function.

Finally, since this callback is emitted once for each missing
intermediate certificate, it's probably less confusing if we talk about
only a single missing intermediate here. Yes, there could be multiple
missing certificates, but a single invocation of this callback can only
deal with one.

Signed-off-by: Michael Catanzaro <mcatanzaro@gnome.org>