build: do not install .hmac files
It turned out that distro package building process might perform
post-processing (e.g., strip) of the shared libraries after install,
and that may cause inconsistency with the installed .hmac files.

Let's not try too hard on this but defer the final hmac calculation to
distributions.  It is still useful to keep our own fipshmac as it
makes it easier to run FIPS tests.

Signed-off-by: Daiki Ueno <ueno@gnu.org>