certtool: order DN components by scale.
DN components are expected to be ordered by scale, with the wire format
representing larger-scale components (like country or organization) before
smaller-scale components (like state or organizationalUnit).

The bulk of the changes here of course are changes to the target
certificates in the test suite.

Note that a change was necessary in tests/cert-tests/crq.sh because it
tests the "interactive" mode of certtool.  If any user is scripting
certtool in this way, this change will cause a backwards-incompatible
break.  However, I think this is OK -- the supported scripted/batch
mode for certtool should use a template file, and I don't think it's
important to maintain a strict api on the interactive mode.

The main change here is to order the DN from least-specific-to-most,
in particular:

    country, state, locality, org, orgunit, cn, uid

But I've also made an additional arbitrary choice, which is that DC
(domain component) comes *after* uid.  This was already the case in
certificate generation, but in *request* generation, it was the other
way around.  I've changed request generation to match this ordering
from certificate generation.

Closes: #1243

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>